Moved Flow extensions to common.extension.flow and separated them into different files
