+++ client/service/remote/src/commonMain/kotlin/com/thebrownfoxx/neon/client/remote/service/RemoteGroupManager.kt
@@ -9,7 +9,7 @@ import com.thebrownfoxx.neon.client.service.GroupManager.GetMembersError
 import com.thebrownfoxx.neon.client.websocket.WebSocketSubscriber
 import com.thebrownfoxx.neon.client.websocket.subscribeAsFlow
 import com.thebrownfoxx.neon.common.data.Cache
-import com.thebrownfoxx.neon.common.extension.mirrorTo
+import com.thebrownfoxx.neon.common.extension.flow.mirrorTo
 import com.thebrownfoxx.neon.common.type.id.GroupId
 import com.thebrownfoxx.neon.common.type.id.MemberId
 import com.thebrownfoxx.neon.server.route.websocket.group.GetGroupMembersGroupNotFound
