+++ common/src/commonMain/kotlin/com/thebrownfoxx/neon/common/extension/flow/Flow.kt
@@ -1,0 +1,5 @@
+package com.thebrownfoxx.neon.common.extension.flow
+
+import kotlinx.coroutines.flow.flowOf
+
+fun <T> T.flow() = flowOf(this)
